fix(src): rename folder for npm link - I62
