+++ src/FormattingToolbar/index.js
