+++ src/FormattingToolbar/toolbar.css
