+++ src/FormattingToolbar/toolbarMethods.js
