+++ src/FormattingToolbar/toolbarStyles.js
