+++ src/FormattingToolbar/toolbarTooltip.js
